Remove Purescript because broken
As with all things Nix, the layers upon layers of abstractions end up
being a nightmare to maintain. I upgraded Nixpkgs, got a newer PS
version, that was incompatible with some dependencies where FFI used
CommonJS. So I upgraded the dependencies, but that gave me a new
upstream Purescript whatever URL for the package set and then I got
errors about Nix remote calls in the sandbox... screw it.