Store attachments in SQLite
This commit is backwards compatible. All newly uploaded attachments are
stored in SQLite. When an event is deleted, we try to delete attachments
both in SQLite and on disk, while always checking if file paths exist.

The attachment middleware still serves attachment content. For any given
path it tries to get the attachment from SQLite. If it finds it and it
has content, it serves it. If it finds it but there's no content, it
tries to read the content from disk.