Move attachments to SQLite
Wow! It is SUCH A PAIN to work with this code base. I blame myself, of
course, but Haskell really isn't helping.

Try moving the Event.Id module into Event.Event. You'll immediately get
cyclic import errors. I'll just leave it the way it is for now.

Furthermore, this code base has too much sharing that's not well thought
through. It's a textbook example of wrong abstractions.

This commit:
* Moves attachments into SQLite and removes the existing Attachment.*
  modules
* Cleans up most, but not all, attachment file storage related code
* Simplifies a few things in the Events.* modules by inlining and
  duplicating code that has minor modifications for the various use
  cases. Example: if you parse the parameters when updating an event,
  you need to handle the case of the pre-existing files. The same is not
  true when creating an event.

I'll simplifying things in this manner.