Remove internal state from App.Env
Having the internal state creating outside of the server means it's
effectively not temporary. It'll live for as long as the app lives.

I think I can safely move this internal state into each request that
needs it, now that I refactored how attachments works.

1. /GET edit event
2. Fill out form and pick files
3. Hit save
4. Uploads are read into InternalState backend
5. If form is valid, attachments are moved to SQLite
6. Clean up the files

5. If form is invalid, don't store attachments, return errors
6. Clean up the files

I was thinking if I could avoid re-uploading attachments but I don't
think that's how it works, regardless of what I do on the server. REST
is stateless after all. Also, the files here are so small that it
doesn't matter anyway. Uploading a 100kb .pdf three times is fine.